requirements.txt
torch>=1.13.1
tqdm
numpy>=1.17.0
imageio==2.6.1
gym==0.18.0
attrs>=21.4.0
hydra-core==1.3.2
omegaconf==2.3.0
d4rl==1.1
mujoco==2.3.6
